Updates deps, versions, drops gulp etc.
This is a HUGE update. With latest deps for the projects it lays
down, and the CLI tool itself.

We are dropping gulp completely and switching to npm scripts for
both the project and the CLI tool which frees us from more deps and
makes life much much nicer and faster.

Also updates all of the dependencies to latest and gets things in
ship shape.

Lastly it moves to monument 2.4.0+ which brings all the latest
features to the CLI