Updates the connectionStub and other things
ConnectionStub which is the backbone of route testing
was broken by the update to harken in monument 3.0.11.
This patch fixes it.

It also bumps the version of monument being used, updates
the default name to the name of the current version
name of monument, updates the test.sh file with some
logging...

and that's pretty much it